Add files via upload
Changed data types of all id's, client's cash and shop's income fields. Created new Product entity and made connection between this and ProductData entity. Created repositories for all entities and services interfaces. Added lombok dependency to project and replaced getters, setters, no argument constructors and toString methods with its annotations.

+++ src/main/java/com/serhiihurin/shop/online_shop/services/ClientService.java
@@ -9,7 +9,7 @@ public interface ClientService {
 
     void saveClient(Client client);
 
-    Client getClient(int id);
+    Client getClient(Long id);
 
-    void deleteClient(int id);
+    void deleteClient(Long id);
 }
